requirements-rpi.txt
# Inference (ARM64 kompatibilno s Python 3.11)
onnxruntime>=1.25.0
rpi-lgpio>=0.6
picamera2>=0.3.34
Pillow>=11.1.0

# Web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Data
numpy>=2.2.4
pandas>=2.0.0

# Utils
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0
imutils>=0.5.4
